requirements.txt
django
djangorestframework
django-rest-knox
psycopg2-binary
python-dotenv